# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vine Server API
  version: 1.0.0
  description: API documentation for authentication, wine cards, comments, ratings, and favorites.
  license:
    name: ISC
    url: https://opensource.org/license/isc-license-txt
servers:
  - url: https://api.vine-server.local
    description: API server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CardIdParam:
      name: id
      in: path
      required: true
      description: Card ID (MongoDB ObjectId)
      schema:
        type: string
        pattern: '^[a-fA-F0-9]{24}$'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - message
...
    Owner:
      type: object
      properties:
        _id:
          type: string
          example: 65f2d95a6f7d0e4f7f8f2d11
        name:
          type: string
          example: Yurii
        email:
          type: string
          format: email
          example: user@example.com
    RatingItem:
      type: object
      properties:
        _id:
          type: string
        userId:
          oneOf:
            - type: string
              example: 65f2d95a6f7d0e4f7f8f2d11
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
        username:
          type: string
          example: Yurii
        value:
          type: number
          minimum: 0
          maximum: 10
          example: 8.5
    CommentItem:
      type: object
      properties:
        _id:
          type: string
          example: 65f2e4bcd4a4f6f48d839acd
        userId:
          type: string
          example: 65f2d95a6f7d0e4f7f8f2d11
        username:
          type: string
          example: Yurii
        text:
          type: string
          maxLength: 1000
          example: Nice taste and aroma.
        createdAt:
          type: string
          format: date-time
          example: '2026-02-15T12:00:00.000Z'
    CardBaseFields:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
          example: Prosecco DOC
        color:
          type: string
          enum: [bianco, rosso, rosato, light, dark]
          example: bianco
        type:
          type: string
          enum:
            [wine, beer, whiskey, brandy, gin, rum, liqueur, grappa, vodka, other]
          example: whiskey
        sweetness:
          type: string
          enum: [secco, abboccato, amabile, dolce]
          description: Wine sweetness/style, only used for wine cards.
          example: secco
        alcohol:
          type: number
          minimum: 0
          maximum: 100
          example: 11.5
        volume:
          type: number
          minimum: 1
          maximum: 100000
          description: Drink volume in milliliters
          example: 750
        winery:
          type: string
          minLength: 2
          maxLength: 120
          example: Villa Sandi
        region:
          type: string
          minLength: 2
          maxLength: 120
          example: Veneto
        country:
          type: string
          minLength: 2
          maxLength: 120
          example: Italy
        anno:
          type: integer
          minimum: 1900
          maximum: 2030
          example: 2022
        price:
          type: number
          minimum: 0
          example: 15.99
        frizzante:
          type: boolean
          example: true
        unfiltered:
          type: boolean
          example: false
        description:
          type: string
          maxLength: 2000
          example: Fresh and fruity sparkling wine.
    Card:
      allOf:
        - $ref: '#/components/schemas/CardBaseFields'
        - type: object
          properties:
            _id:
              type: string
              example: 65f2e2d6b8e6ce4f3c1d9f0a
            img:
              type: string
              format: uri
              example: https://res.cloudinary.com/demo/image/upload/wines/default.jpg
            rating:
              type: number
              minimum: 0
              maximum: 10
              example: 8.3
            ratings:
              type: array
              items:
                $ref: '#/components/schemas/RatingItem'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentItem'
            owner:
              oneOf:
                - type: string
                - $ref: '#/components/schemas/Owner'
            favorites:
              type: array
              items:
                type: string
            isFavorite:
              type: boolean
              example: false
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    RateCardRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          type: number
          minimum: 0
          maximum: 10
          example: 9
        username:
          type: string
          maxLength: 100
          example: Yurii
...
paths:
...
  /cards/{id}/rate:
    patch:
      tags: [Cards]
      summary: Add or update card rating
      operationId: rateCard
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CardIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateCardRequest'
      responses:
        '200':
          description: Updated card with new rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...